Code Quality & Documentation Overhaul
