+++ convenience.go
@@ -1,6 +1,8 @@
 package jwt
 
 import (
+	"crypto/sha256"
+	"encoding/hex"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -22,72 +24,55 @@ var cache = &processorCache{
 	entries: make(map[string]*cacheEntry, 16),
 }
 
-// CreateToken creates a JWT token using a cached processor.
-// This is a convenience function for simple use cases without rate limiting.
-// For production environments with rate limiting, use the Processor API.
-// The secret key must be at least 32 bytes long.
-func CreateToken(secretKey string, claims Claims) (string, error) {
-	if len(secretKey) < 32 {
-		return "", ErrInvalidSecretKey
-	}
+func hashSecretKey(secretKey string) string {
+	hash := sha256.Sum256([]byte(secretKey))
+	return hex.EncodeToString(hash[:])
+}
 
+func CreateToken(secretKey string, claims Claims) (string, error) {
 	processor, release, err := getProcessor(secretKey)
 	if err != nil {
 		return "", err
 	}
 	defer release()
-
 	return processor.CreateToken(claims)
 }
 
-// ValidateToken validates a JWT token using a cached processor.
-// This is a convenience function for simple use cases without rate limiting.
-// For production environments with rate limiting, use the Processor API.
-// The secret key must be at least 32 bytes long.
 func ValidateToken(secretKey, tokenString string) (Claims, bool, error) {
-	if len(secretKey) < 32 {
-		return Claims{}, false, ErrInvalidSecretKey
-	}
-
 	processor, release, err := getProcessor(secretKey)
 	if err != nil {
 		return Claims{}, false, err
 	}
 	defer release()
-
 	return processor.ValidateToken(tokenString)
 }
 
-// RevokeToken revokes a JWT token using a cached processor.
-// This is a convenience function for simple use cases without rate limiting.
-// For production environments with rate limiting, use the Processor API.
-// The secret key must be at least 32 bytes long.
 func RevokeToken(secretKey, tokenString string) error {
-	if len(secretKey) < 32 {
-		return ErrInvalidSecretKey
-	}
-
 	processor, release, err := getProcessor(secretKey)
 	if err != nil {
 		return err
 	}
 	defer release()
-
 	return processor.RevokeToken(tokenString)
 }
 
 func getProcessor(secretKey string) (*Processor, func(), error) {
+	if len(secretKey) < 32 {
+		return nil, func() {}, ErrInvalidSecretKey
+	}
+
+	keyHash := hashSecretKey(secretKey)
 	now := time.Now().Unix()
 
 	cache.mu.RLock()
-	entry, exists := cache.entries[secretKey]
-	cache.mu.RUnlock()
-
+	entry, exists := cache.entries[keyHash]
 	if exists {
 		entry.lastAccess.Store(now)
 		entry.refCount.Add(1)
+		cache.mu.RUnlock()
 		return entry.processor, func() { entry.refCount.Add(-1) }, nil
 	}
+	cache.mu.RUnlock()
 
 	processor, err := New(secretKey)
 	if err != nil {
@@ -97,14 +82,13 @@ func getProcessor(secretKey string) (*Processor, func(), error) {
 	cache.mu.Lock()
 	defer cache.mu.Unlock()
 
-	if entry, exists := cache.entries[secretKey]; exists {
+	if entry, exists := cache.entries[keyHash]; exists {
 		processor.Close()
 		entry.lastAccess.Store(now)
 		entry.refCount.Add(1)
 		return entry.processor, func() { entry.refCount.Add(-1) }, nil
 	}
 
-	const maxCacheSize = 100
 	if len(cache.entries) >= maxCacheSize {
 		evictOldestUnsafe()
 	}
@@ -112,9 +96,11 @@ func getProcessor(secretKey string) (*Processor, func(), error) {
 	entry = &cacheEntry{processor: processor}
 	entry.lastAccess.Store(now)
 	entry.refCount.Store(1)
-	cache.entries[secretKey] = entry
+	cache.entries[keyHash] = entry
 
-	cleanupCacheIfNeeded(now)
+	if now-cache.lastCleanup.Load() >= cacheCleanupInterval {
+		go cleanupCacheAsync(now)
+	}
 
 	return processor, func() { entry.refCount.Add(-1) }, nil
 }
@@ -124,69 +110,84 @@ func evictOldestUnsafe() {
 		return
 	}
 
-	oldestKey := ""
+	var oldestKey string
 	oldestTime := int64(1<<63 - 1)
 
 	for k, entry := range cache.entries {
 		if entry.refCount.Load() > 0 {
 			continue
 		}
-		lastAccess := entry.lastAccess.Load()
-		if lastAccess < oldestTime {
+		if lastAccess := entry.lastAccess.Load(); lastAccess < oldestTime {
 			oldestKey = k
 			oldestTime = lastAccess
 		}
 	}
 
 	if oldestKey != "" {
-		if entry, exists := cache.entries[oldestKey]; exists {
-			if entry.processor != nil {
-				entry.processor.Close()
-			}
-			delete(cache.entries, oldestKey)
+		if entry := cache.entries[oldestKey]; entry != nil && entry.processor != nil {
+			entry.processor.Close()
 		}
+		delete(cache.entries, oldestKey)
 	}
 }
 
 const (
+	maxCacheSize         = 100
 	cacheCleanupInterval = 300  // 5 minutes in seconds
 	cacheMaxIdleTime     = 3600 // 1 hour in seconds
+	batchEvictSize       = 5    // Evict multiple entries when cache is full
 )
 
-func cleanupCacheIfNeeded(now int64) {
-	lastCleanup := cache.lastCleanup.Load()
-	if now-lastCleanup < cacheCleanupInterval {
+func cleanupCacheAsync(now int64) {
+	if !cache.lastCleanup.CompareAndSwap(cache.lastCleanup.Load(), now) {
 		return
 	}
 
-	if !cache.lastCleanup.CompareAndSwap(lastCleanup, now) {
-		return
-	}
+	cache.mu.Lock()
+	defer cache.mu.Unlock()
 
+	toDelete := make([]string, 0, len(cache.entries)/10)
 	for key, entry := range cache.entries {
 		if entry.refCount.Load() > 0 {
 			continue
 		}
 		if now-entry.lastAccess.Load() > cacheMaxIdleTime {
-			if entry.processor != nil {
-				entry.processor.Close()
-			}
-			delete(cache.entries, key)
+			toDelete = append(toDelete, key)
 		}
 	}
-}
 
-// ClearCache clears all cached processors and releases their resources.
-// This is primarily useful for testing and graceful shutdown scenarios.
-// After calling ClearCache, subsequent calls to convenience functions will create new processors.
-func ClearCache() {
-	cache.mu.Lock()
-	defer cache.mu.Unlock()
-
-	for key, entry := range cache.entries {
-		if entry.processor != nil {
+	for _, key := range toDelete {
+		if entry := cache.entries[key]; entry != nil && entry.processor != nil {
 			entry.processor.Close()
 		}
 		delete(cache.entries, key)
 	}
+}
+
+type CacheStats struct {
+	Size        int
+	MaxSize     int
+	LastCleanup int64
+}
+
+func GetCacheStats() CacheStats {
+	cache.mu.RLock()
+	defer cache.mu.RUnlock()
+	return CacheStats{
+		Size:        len(cache.entries),
+		MaxSize:     maxCacheSize,
+		LastCleanup: cache.lastCleanup.Load(),
+	}
+}
+
+func ClearCache() {
+	cache.mu.Lock()
+	defer cache.mu.Unlock()
+
+	for _, entry := range cache.entries {
+		if entry != nil && entry.processor != nil {
+			entry.processor.Close()
+		}
+	}
+	clear(cache.entries)
 }
