+++ internal/manager.go
@@ -1,24 +1,23 @@
-package blacklist
+package internal
 
 import (
 	"fmt"
 	"time"
-
-	"github.com/cybergodev/jwt/internal/core"
 )
 
-// Manager wraps a Store and provides high-level token blacklist operations.
-// It handles token parsing and expiration management.
+type tokenClaims struct {
+	ID        string `json:"jti,omitempty"`
+	ExpiresAt int64  `json:"exp,omitempty"`
+}
+
 type Manager struct {
 	store Store
 }
 
-// NewManager creates a new blacklist manager with the given store
 func NewManager(store Store) *Manager {
 	return &Manager{store: store}
 }
 
-// BlacklistToken adds a token to the blacklist
 func (m *Manager) BlacklistToken(tokenID string, expiresAt time.Time) error {
 	if tokenID == "" {
 		return fmt.Errorf("token ID cannot be empty")
@@ -26,7 +25,6 @@ func (m *Manager) BlacklistToken(tokenID string, expiresAt time.Time) error {
 	return m.store.Add(tokenID, expiresAt)
 }
 
-// IsBlacklisted checks if a token is blacklisted
 func (m *Manager) IsBlacklisted(tokenID string) (bool, error) {
 	if tokenID == "" {
 		return false, nil
@@ -34,20 +32,13 @@ func (m *Manager) IsBlacklisted(tokenID string) (bool, error) {
 	return m.store.Contains(tokenID)
 }
 
-// BlacklistTokenString extracts token ID from token string and blacklists it
 func (m *Manager) BlacklistTokenString(tokenString string) error {
 	if tokenString == "" {
 		return fmt.Errorf("token string cannot be empty")
 	}
 
-	type minimalClaims struct {
-		ID        string `json:"jti,omitempty"`
-		ExpiresAt int64  `json:"exp,omitempty"`
-	}
-
-	claims := &minimalClaims{}
-
-	_, _, err := core.ParseUnverified(tokenString, claims)
+	claims := &tokenClaims{}
+	_, _, err := ParseUnverified(tokenString, claims)
 	if err != nil {
 		return fmt.Errorf("failed to parse token: %w", err)
 	}
@@ -56,7 +47,7 @@ func (m *Manager) BlacklistTokenString(tokenString string) error {
 		return fmt.Errorf("token does not contain a valid ID (jti)")
 	}
 
-	expiresAt := time.Now().Add(24 * time.Hour)
+	expiresAt := time.Now().Add(7 * 24 * time.Hour)
 	if claims.ExpiresAt > 0 {
 		expiresAt = time.Unix(claims.ExpiresAt, 0)
 	}
@@ -64,7 +55,6 @@ func (m *Manager) BlacklistTokenString(tokenString string) error {
 	return m.BlacklistToken(claims.ID, expiresAt)
 }
 
-// Close closes the manager and underlying store
 func (m *Manager) Close() error {
 	return m.store.Close()
 }
