+++ processor.go
@@ -6,10 +6,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/cybergodev/jwt/internal/blacklist"
-	"github.com/cybergodev/jwt/internal/core"
-	"github.com/cybergodev/jwt/internal/security"
-	"github.com/cybergodev/jwt/internal/signing"
+	"github.com/cybergodev/jwt/internal"
 )
 
 type Processor struct {
@@ -18,7 +15,7 @@ type Processor struct {
 	refreshTokenTTL  time.Duration
 	issuer           string
 	signingMethod    SigningMethod
-	blacklistManager *blacklist.Manager
+	blacklistManager *internal.Manager
 	rateLimiter      *RateLimiter
 	closed           atomic.Bool
 }
@@ -49,19 +46,10 @@ func newProcessor(secretKey string, blacklistConfig BlacklistConfig, config ...C
 	var cfg Config
 	if len(config) > 0 {
 		cfg = config[0]
-		cfg.SecretKey = secretKey
 	} else {
 		cfg = DefaultConfig()
-		cfg.SecretKey = secretKey
 	}
-
-	if cfg.SigningMethod == "" {
-		cfg.SigningMethod = SigningMethodHS256
-	}
-
-	if cfg.Issuer == "" {
-		cfg.Issuer = "jwt-service"
-	}
+	cfg.SecretKey = secretKey
 
 	if err := cfg.Validate(); err != nil {
 		return nil, err
@@ -70,44 +58,30 @@ func newProcessor(secretKey string, blacklistConfig BlacklistConfig, config ...C
 	secretKeyBytes := make([]byte, len(cfg.SecretKey))
 	copy(secretKeyBytes, cfg.SecretKey)
 
-	store := blacklist.NewStore(blacklist.Config{
-		CleanupInterval:   blacklistConfig.CleanupInterval,
-		MaxSize:           blacklistConfig.MaxSize,
-		EnableAutoCleanup: blacklistConfig.EnableAutoCleanup,
-	})
-	blacklistMgr := blacklist.NewManager(store)
+	store := internal.NewMemoryStore(
+		blacklistConfig.MaxSize,
+		blacklistConfig.CleanupInterval,
+		blacklistConfig.EnableAutoCleanup,
+	)
 
 	var rateLimiter *RateLimiter
 	if cfg.RateLimiter != nil {
 		rateLimiter = cfg.RateLimiter
 	} else if cfg.EnableRateLimit {
-		rate := cfg.RateLimitRate
-		window := cfg.RateLimitWindow
-		if rate <= 0 {
-			rate = 100
-		}
-		if window <= 0 {
-			window = time.Minute
-		}
-		rateLimiter = NewRateLimiter(rate, window)
+		rateLimiter = NewRateLimiter(cfg.RateLimitRate, cfg.RateLimitWindow)
 	}
 
-	processor := &Processor{
+	return &Processor{
 		secretKey:        secretKeyBytes,
 		accessTokenTTL:   cfg.AccessTokenTTL,
 		refreshTokenTTL:  cfg.RefreshTokenTTL,
 		issuer:           cfg.Issuer,
 		signingMethod:    cfg.SigningMethod,
-		blacklistManager: blacklistMgr,
+		blacklistManager: internal.NewManager(store),
 		rateLimiter:      rateLimiter,
-	}
-
-	return processor, nil
+	}, nil
 }
 
-// CreateToken creates a JWT access token with the provided claims.
-// The token will be signed using the processor's configured signing method.
-// Returns an error if the processor is closed, claims are invalid, or rate limit is exceeded.
 func (p *Processor) CreateToken(claims Claims) (string, error) {
 	if p.closed.Load() {
 		return "", ErrProcessorClosed
@@ -117,13 +91,9 @@ func (p *Processor) CreateToken(claims Claims) (string, error) {
 		return "", err
 	}
 
-	return p.createTokenWithClaims(claims)
+	return p.createTokenWithTTL(claims, p.accessTokenTTL)
 }
 
-// ValidateToken validates a JWT token and returns the claims.
-// Returns the claims, a boolean indicating validity, and an error if validation fails.
-// The boolean is false if the token is expired, not yet valid, or has an invalid issuer.
-// Returns ErrTokenRevoked if the token has been blacklisted.
 func (p *Processor) ValidateToken(tokenString string) (Claims, bool, error) {
 	if p.closed.Load() {
 		return Claims{}, false, ErrProcessorClosed
@@ -135,7 +105,9 @@ func (p *Processor) ValidateToken(tokenString string) (Claims, bool, error) {
 
 	claims, valid, err := p.validateTokenInternal(tokenString)
 	if err != nil {
-		putClaims(claims)
+		if claims != nil {
+			putClaims(claims)
+		}
 		return Claims{}, false, ErrInvalidToken
 	}
 
@@ -145,12 +117,10 @@ func (p *Processor) ValidateToken(tokenString string) (Claims, bool, error) {
 	}
 
 	if claims.ID != "" {
-		isBlacklisted, err := p.blacklistManager.IsBlacklisted(claims.ID)
-		if err != nil {
+		if isBlacklisted, err := p.blacklistManager.IsBlacklisted(claims.ID); err != nil {
 			putClaims(claims)
 			return Claims{}, false, err
-		}
-		if isBlacklisted {
+		} else if isBlacklisted {
 			putClaims(claims)
 			return Claims{}, false, ErrTokenRevoked
 		}
@@ -161,9 +131,6 @@ func (p *Processor) ValidateToken(tokenString string) (Claims, bool, error) {
 	return result, valid, nil
 }
 
-// CreateRefreshToken creates a refresh token with longer TTL.
-// Refresh tokens are used to obtain new access tokens without re-authentication.
-// Returns an error if the processor is closed, claims are invalid, or rate limit is exceeded.
 func (p *Processor) CreateRefreshToken(claims Claims) (string, error) {
 	if p.closed.Load() {
 		return "", ErrProcessorClosed
@@ -176,9 +143,6 @@ func (p *Processor) CreateRefreshToken(claims Claims) (string, error) {
 	return p.createTokenWithTTL(claims, p.refreshTokenTTL)
 }
 
-// RefreshToken validates a refresh token and creates a new access token.
-// The new access token will have the same claims as the refresh token but with a new expiration time.
-// Returns an error if the refresh token is invalid, expired, or the processor is closed.
 func (p *Processor) RefreshToken(refreshTokenString string) (string, error) {
 	if p.closed.Load() {
 		return "", ErrProcessorClosed
@@ -204,13 +168,9 @@ func (p *Processor) RefreshToken(refreshTokenString string) (string, error) {
 	newClaims.ExpiresAt = NumericDate{}
 	newClaims.ID = ""
 
-	return p.createTokenWithClaims(newClaims)
+	return p.createTokenWithTTL(newClaims, p.accessTokenTTL)
 }
 
-// Close gracefully shuts down the processor and securely clears the secret key.
-// This method should always be called when the processor is no longer needed.
-// After calling Close, all subsequent operations will return ErrProcessorClosed.
-// It is safe to call Close multiple times.
 func (p *Processor) Close() error {
 	if !p.closed.CompareAndSwap(false, true) {
 		return ErrProcessorClosed
@@ -229,19 +189,14 @@ func (p *Processor) Close() error {
 	}
 
 	if p.secretKey != nil {
-		security.ZeroBytes(p.secretKey)
+		internal.ZeroBytes(p.secretKey)
 	}
 
 	return closeErr
 }
 
-// IsClosed returns true if the processor has been closed
 func (p *Processor) IsClosed() bool {
 	return p.closed.Load()
-}
-
-func (p *Processor) createTokenWithClaims(claims Claims) (string, error) {
-	return p.createTokenWithTTL(claims, p.accessTokenTTL)
 }
 
 func (p *Processor) createTokenWithTTL(claims Claims, ttl time.Duration) (string, error) {
@@ -265,38 +220,27 @@ func (p *Processor) createTokenWithTTL(claims Claims, ttl time.Duration) (string
 		claimsCopy.Issuer = p.issuer
 	}
 	if claimsCopy.ID == "" {
-		claimsCopy.ID = core.GenerateTokenIDFast()
+		tokenID, err := internal.GenerateTokenIDFast()
+		if err != nil {
+			return "", fmt.Errorf("failed to generate token ID: %w", err)
+		}
+		claimsCopy.ID = tokenID
 	}
 
-	signingMethod, err := signing.GetInternalSigningMethod(string(p.signingMethod))
+	signingMethod, err := internal.GetInternalSigningMethod(string(p.signingMethod))
 	if err != nil {
 		return "", err
 	}
 
-	token := core.NewTokenWithClaims(signingMethod, claimsCopy)
+	token := internal.NewTokenWithClaims(signingMethod, claimsCopy)
 	return token.SignedString(p.secretKey)
 }
 
 func copyClaims(dst, src *Claims) {
-	*dst = Claims{
-		UserID:      src.UserID,
-		Username:    src.Username,
-		Role:        src.Role,
-		SessionID:   src.SessionID,
-		ClientID:    src.ClientID,
-		Permissions: append(dst.Permissions[:0], src.Permissions...),
-		Scopes:      append(dst.Scopes[:0], src.Scopes...),
-		Extra:       dst.Extra,
-		RegisteredClaims: RegisteredClaims{
-			Issuer:    src.Issuer,
-			Subject:   src.Subject,
-			Audience:  append(dst.Audience[:0], src.Audience...),
-			ExpiresAt: src.ExpiresAt,
-			NotBefore: src.NotBefore,
-			IssuedAt:  src.IssuedAt,
-			ID:        src.ID,
-		},
-	}
+	*dst = *src
+	dst.Permissions = append(dst.Permissions[:0], src.Permissions...)
+	dst.Scopes = append(dst.Scopes[:0], src.Scopes...)
+	dst.Audience = append(dst.Audience[:0], src.Audience...)
 
 	if len(src.Extra) > 0 {
 		if dst.Extra == nil {
@@ -313,9 +257,8 @@ func copyClaims(dst, src *Claims) {
 func (p *Processor) validateTokenInternal(tokenString string) (*Claims, bool, error) {
 	claims := getClaims()
 
-	token, err := core.ParseWithClaims(tokenString, claims, func(token *core.Core) (any, error) {
-		alg, ok := token.Header["alg"].(string)
-		if !ok || alg != string(p.signingMethod) {
+	token, err := internal.ParseWithClaims(tokenString, claims, func(token *internal.Core) (any, error) {
+		if alg, ok := token.Header["alg"].(string); !ok || alg != string(p.signingMethod) {
 			return nil, ErrInvalidToken
 		}
 		return p.secretKey, nil
@@ -330,22 +273,15 @@ func (p *Processor) validateTokenInternal(tokenString string) (*Claims, bool, er
 	}
 
 	now := time.Now()
-	if !claims.ExpiresAt.IsZero() && now.After(claims.ExpiresAt.Time) {
-		return claims, false, nil
-	}
-	if !claims.NotBefore.IsZero() && now.Before(claims.NotBefore.Time) {
-		return claims, false, nil
-	}
-	if claims.Issuer != p.issuer {
+	if (!claims.ExpiresAt.IsZero() && now.After(claims.ExpiresAt.Time)) ||
+		(!claims.NotBefore.IsZero() && now.Before(claims.NotBefore.Time)) ||
+		(claims.Issuer != "" && claims.Issuer != p.issuer) {
 		return claims, false, nil
 	}
 
 	return claims, true, nil
 }
 
-// RevokeToken adds a token to the blacklist, preventing its future use.
-// The token will remain blacklisted until it expires naturally.
-// Returns an error if the processor is closed or the token cannot be parsed.
 func (p *Processor) RevokeToken(tokenString string) error {
 	if p.closed.Load() {
 		return ErrProcessorClosed
@@ -358,9 +294,6 @@ func (p *Processor) RevokeToken(tokenString string) error {
 	return p.blacklistManager.BlacklistTokenString(tokenString)
 }
 
-// IsTokenRevoked checks if a token has been revoked (is in the blacklist).
-// Returns true if the token is blacklisted, false otherwise.
-// Returns an error if the processor is closed or the token cannot be parsed.
 func (p *Processor) IsTokenRevoked(tokenString string) (bool, error) {
 	if p.closed.Load() {
 		return false, ErrProcessorClosed
@@ -370,13 +303,12 @@ func (p *Processor) IsTokenRevoked(tokenString string) (bool, error) {
 		return false, ErrEmptyToken
 	}
 
-	type minimalClaims struct {
+	type idClaims struct {
 		ID string `json:"jti,omitempty"`
 	}
 
-	claims := &minimalClaims{}
-
-	_, _, err := core.ParseUnverified(tokenString, claims)
+	claims := &idClaims{}
+	_, _, err := internal.ParseUnverified(tokenString, claims)
 	if err != nil {
 		return false, fmt.Errorf("failed to parse token: %w", err)
 	}
