+++ ratelimit.go
@@ -5,9 +5,7 @@ import (
 	"time"
 )
 
-// RateLimiter provides rate limiting for JWT operations to prevent abuse.
-// It uses a token bucket algorithm with per-key rate limiting.
-// The rate limiter is thread-safe and can be used concurrently.
+// RateLimiter provides rate limiting for JWT operations using token bucket algorithm.
 type RateLimiter struct {
 	mu         sync.Mutex
 	buckets    map[string]*bucket
@@ -17,16 +15,12 @@ type RateLimiter struct {
 	closed     bool
 }
 
-// bucket represents a token bucket for rate limiting
 type bucket struct {
 	tokens     int
 	lastRefill int64
 }
 
 // NewRateLimiter creates a new rate limiter with the specified rate and window.
-// maxRate is the maximum number of requests allowed per window.
-// window is the time window for rate limiting (e.g., time.Minute).
-// If maxRate or window is invalid, sensible defaults are used.
 func NewRateLimiter(maxRate int, window time.Duration) *RateLimiter {
 	if maxRate <= 0 {
 		maxRate = 100
@@ -44,17 +38,16 @@ func NewRateLimiter(maxRate int, window time.Duration) *RateLimiter {
 }
 
 // Allow checks if a single request is allowed for the given key.
-// Returns true if the request is allowed, false if rate limit is exceeded.
-// An empty key always returns false.
 func (rl *RateLimiter) Allow(key string) bool {
 	return rl.AllowN(key, 1)
 }
 
 // AllowN checks if n requests are allowed for the given key.
-// Returns true if all n requests are allowed, false if rate limit would be exceeded.
-// An empty key always returns false. n <= 0 always returns true.
 func (rl *RateLimiter) AllowN(key string, n int) bool {
-	if n <= 0 {
+	if n < 0 {
+		return false
+	}
+	if n == 0 {
 		return true
 	}
 	if key == "" {
@@ -86,12 +79,13 @@ func (rl *RateLimiter) AllowN(key string, n int) bool {
 	}
 
 	elapsed := nowNano - b.lastRefill
+	windowNano := int64(rl.window)
 
-	if elapsed >= int64(rl.window) {
+	if elapsed >= windowNano {
 		b.tokens = rl.maxRate
 		b.lastRefill = nowNano
 	} else if elapsed > 0 {
-		tokensToAdd := int(float64(rl.maxRate) * float64(elapsed) / float64(rl.window))
+		tokensToAdd := int((int64(rl.maxRate) * elapsed) / windowNano)
 		if tokensToAdd > 0 {
 			b.tokens += tokensToAdd
 			if b.tokens > rl.maxRate {
@@ -109,8 +103,7 @@ func (rl *RateLimiter) AllowN(key string, n int) bool {
 	return false
 }
 
-// Reset removes the rate limit bucket for the given key, effectively resetting its rate limit.
-// This is useful for clearing rate limits after successful authentication or for testing.
+// Reset removes the rate limit bucket for the given key.
 func (rl *RateLimiter) Reset(key string) {
 	if key == "" {
 		return
@@ -122,8 +115,6 @@ func (rl *RateLimiter) Reset(key string) {
 }
 
 // Close closes the rate limiter and releases all resources.
-// After calling Close, all subsequent Allow/AllowN calls will return false.
-// It is safe to call Close multiple times.
 func (rl *RateLimiter) Close() {
 	rl.mu.Lock()
 	defer rl.mu.Unlock()
@@ -142,7 +133,7 @@ func (rl *RateLimiter) evictOldestUnsafe() {
 		return
 	}
 
-	oldestKey := ""
+	var oldestKey string
 	oldestTime := int64(1<<63 - 1)
 
 	for key, b := range rl.buckets {
