+++ time.go
@@ -5,27 +5,27 @@ import (
 	"time"
 )
 
-// NumericDate represents a JSON numeric date value as specified in RFC 7519.
-// It stores time as Unix timestamp (seconds since epoch) for JWT compatibility.
 type NumericDate struct {
 	time.Time
 }
 
-// NewNumericDate creates a new NumericDate from time.Time
 func NewNumericDate(t time.Time) NumericDate {
 	return NumericDate{Time: t}
 }
 
-// MarshalJSON implements json.Marshaler interface
 func (date *NumericDate) MarshalJSON() ([]byte, error) {
 	if date.Time.IsZero() {
 		return []byte("null"), nil
 	}
 
-	return fmt.Appendf(nil, "%d", date.Unix()), nil
+	unix := date.Unix()
+	if unix < 0 || unix > 253402300799 {
+		return []byte("null"), nil
+	}
+
+	return fmt.Appendf(nil, "%d", unix), nil
 }
 
-// UnmarshalJSON implements json.Unmarshaler interface
 func (date *NumericDate) UnmarshalJSON(b []byte) error {
 	if len(b) == 0 || string(b) == "null" {
 		date.Time = time.Time{}
@@ -33,23 +33,24 @@ func (date *NumericDate) UnmarshalJSON(b []byte) error {
 	}
 
 	s := string(b)
-	if s[0] == '"' && s[len(s)-1] == '"' {
+	if len(s) >= 2 && s[0] == '"' && s[len(s)-1] == '"' {
 		s = s[1 : len(s)-1]
 	}
 
-	if s == "" {
+	if s == "" || s == "null" {
 		date.Time = time.Time{}
 		return nil
 	}
 
 	var unix int64
-	if _, err := fmt.Sscanf(s, "%d", &unix); err == nil {
-		if unix < 0 || unix > 253402300799 {
-			return fmt.Errorf("invalid unix timestamp: %d", unix)
-		}
-		date.Time = time.Unix(unix, 0).UTC()
-		return nil
+	if _, err := fmt.Sscanf(s, "%d", &unix); err != nil {
+		return fmt.Errorf("invalid time format: expected unix timestamp, got %s", s)
 	}
 
-	return fmt.Errorf("invalid time format: expected unix timestamp, got %s", s)
+	if unix < 0 || unix > 253402300799 {
+		return fmt.Errorf("invalid unix timestamp: %d", unix)
+	}
+
+	date.Time = time.Unix(unix, 0).UTC()
+	return nil
 }
